% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/honeylogging-package.R
\docType{package}
\name{honeylogging}
\alias{honeylogging}
\alias{honeylogging-package}
\title{Provide R integration with honeybadger.}
\description{
Provide R integration with honeybadger.
}

