% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log.R
\name{with_logging}
\alias{with_logging}
\title{Wrap an expression with this function, and on error it will POST
error stacktrace to honeybadger}
\usage{
with_logging(expr, message = "NA", tags = list())
}
\arguments{
\item{expr}{expression. Expression to be evaluated.}

\item{message}{character. Message to be displayed on henoeybadger for this error.}

\item{tags}{list. Additional tags for sorting on honeybadger UI.}
}
\description{
Wrap an expression with this function, and on error it will POST
error stacktrace to honeybadger
}

